refactor: centralize expiry check
